+++ spec/utils/cordova.spec.js
@@ -422,4 +422,3 @@ describe('filterArgumentsForCordova', function() {
     expect(resultArgs).toEqual(['build', '--target="ios"']);
   });
 });
-